Crate mp4ameta

A library for reading and writing iTunes style MPEG-4 audio metadata.

Example

let mut tag = mp4ameta::Tag::read_from_path("music.m4a").unwrap();

println!("{}", tag.artist().unwrap());

tag.set_artist("artist");

tag.write_to_path("music.m4a").unwrap();

Modules

atom
Contains constants, structs and functions for working with MPEG-4 metadata atoms.

data
Contains constants, structs and functions for working with data held inside MPEG-4 metadata atoms.

types
Contains structs and constants for working with types held inside data atoms.

Structs

Atom
A struct that represents a MPEG-4 audio metadata atom.

AtomData
A struct representing data that is associated with an Atom identifier.

AtomT
A template representing a MPEG-4 audio metadata atom.

DataT
A template used for parsing data defined by Table 3-5 Well-known data types.

Error
A structure able to represent any error that may occur while performing metadata operations.

Ident
A 4 byte atom identifier.

Tag
A MPEG-4 audio tag containing metadata atoms

Enums

AdvisoryRating
An enum describing the rating of a file stored in the rtng atom.

Content
An enum representing the different types of content an atom might have.

ContentT
A template representing the different types of content an atom template might have.

Data
An enum that holds different types of data defined by Table 3-5 Well-known data types.

ErrorKind
Kinds of errors that may occur while performing metadata operations.

MediaType
An enum describing the media type of a file stored in the stik atom.

Constants

STANDARD_GENRES
A list of standard genre codes and values found in the gnre atom. This list is equal to the ID3v1 genre list but all codes are incremented by 1.

Type Definitions

Result
Type alias for the result of tag operations.