Module mp4ameta::atom

Contains constants, structs and functions for working with MPEG-4 metadata atoms.

Structs

Atom
A struct that represents a MPEG-4 audio metadata atom.

AtomData
A struct representing data that is associated with an Atom identifier.

AtomT
A template representing a MPEG-4 audio metadata atom.

FILETYPE_ATOM_T
Lazily initialized static reference to a ftyp atom template.

ITEM_LIST_ATOM_T
Lazily initialized static reference to an atom hierarchy template leading to an empty ilst atom.

Ident
A 4 byte atom identifier.

METADATA_ATOM_T
Lazily initialized static reference to an atom metadata hierarchy template needed to parse metadata.

Constants

ADVISORY_RATING
(rtng)

ALBUM
(©alb)

ALBUM_ARTIST
(aART)

ARTIST
(©ART)

ARTWORK
(covr)

BPM
(tmpo)

CATEGORY
(catg)

COMMENT
(©cmt)

COMPILATION
(cpil)

COMPOSER
(©wrt)

COPYRIGHT
(cprt)

CUSTOM_GENRE
(©gen)

DATA
(data) Identifier of an atom containing typed data.

DESCRIPTION
(desc)

DISC_NUMBER
(disk)

ENCODER
(©too)

FILETYPE
(ftyp) Identifier of an atom information about the filetype.

GAPLESS_PLAYBACK
(pgap)

GROUPING
(©grp)

ITEM_LIST
(ilst) Identifier of an atom containing a list of metadata atoms.

KEYWORD
(keyw)

LYRICS
(©lyr)

MEDIA
(mdia) Identifier of an atom containing information about a tracks media type and data.

MEDIA_HEADER
(mdhd) Identifier of an atom specifying the characteristics of a media atom.

MEDIA_TYPE
(stik)

METADATA
(meta) Identifier of an atom containing a metadata item list.

MOVEMENT
(©mvn)

MOVEMENT_COUNT
(©mvc)

MOVEMENT_INDEX
(©mvi)

MOVIE
(moov) Identifier of an atom containing a structure of children storing metadata.

PODCAST
(pcst)

PODCAST_EPISODE_GLOBAL_UNIQUE_ID
(egid)

PODCAST_URL
(purl)

PURCHASE_DATE
(purd)

SHOW_MOVEMENT
(shwm)

STANDARD_GENRE
(gnre)

TITLE
(©nam)

TRACK
(trak) Identifier of an atom containing information about a single track.

TRACK_NUMBER
(trkn)

TV_EPISODE
(tves)

TV_EPISODE_NUMBER
(tven)

TV_NETWORK_NAME
(tvnn)

TV_SEASON
(tvsn)

TV_SHOW_NAME
(tvsh)

USER_DATA
(udta) Identifier of an atom containing user metadata.

VALID_FILETYPES
A list of valid file types defined by the ftyp atom.

WORK
(©wrk)

YEAR
(©day)

Functions

dump_tag_to
Attempts to dump the metadata atoms to the writer. This doesn't include a complete MPEG-4 container hierarchy and won't result in a usable file.

parse_atoms
Attempts to parse the list of atoms, matching the templates, from the reader.

parse_head
Attempts to parse the atom's head containing a 32 bit unsigned integer determining the size of the atom in bytes and the following 4 byte identifier from the reader.

read_tag_from
Attempts to read MPEG-4 audio metadata from the reader.

write_tag_to
Attempts to write the metadata atoms to the file inside the item list atom.