Module mp4ameta::types

Contains structs and constants for working with types held inside data atoms.

Enums

AdvisoryRating
An enum describing the rating of a file stored in the rtng atom.

MediaType
An enum describing the media type of a file stored in the stik atom.

Constants

AUDIOBOOK
A media type code stored in the stik atom.

BOOKLET
A media type code stored in the stik atom.

CLEAN
An advisory rating code stored in the rtng atom.

INOFFENSIVE
An advisory rating code stored in the rtng atom.

MOVIE
A media type code stored in the stik atom.

MUSIC_VIDEO
A media type code stored in the stik atom.

NORMAL
A media type code stored in the stik atom.

SHORT_FILM
A media type code stored in the stik atom.

TV_SHOW
A media type code stored in the stik atom.

WHACKED_BOOKMARK
A media type code stored in the stik atom.